# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ApoTrack API",
        "description": "Dokumentasi API resmi untuk platform ApoTrack — sistem manajemen apotek digital. API ini digunakan oleh aplikasi mobile Flutter untuk Customer dan Staff Apotek.",
        "contact": {
            "name": "Tim Pengembang ApoTrack",
            "email": "dev@apotrack.id"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Server Aktif (Local Development)"
        }
    ],
    "paths": {
...
        "/api/staff/orders/{id}/ship": {
            "post": {
                "tags": [
                    "Orders (Staff)"
                ],
                "summary": "Minta penjemputan paket kurir (Biteship)",
                "description": "Membuat pesanan pengiriman (pickup request) ke Biteship untuk pesanan dengan layanan DELIVERY yang sudah siap dikirim.",
                "operationId": "staffShipOrder",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID pesanan (UUID)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "019e2ac0-684e-7139-8cb9-c150e027c872"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Data kurir pengiriman",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "courier_code",
                                    "courier_service"
                                ],
                                "properties": {
                                    "courier_code": {
                                        "type": "string",
                                        "example": "jne"
                                    },
                                    "courier_service": {
                                        "type": "string",
                                        "example": "reg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Permintaan pickup berhasil dikirim ke Biteship",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Permintaan pengiriman berhasil dikirim ke Biteship"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Tidak terautentikasi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Gagal memproses pengiriman (misal: Biteship API error atau pesanan belum lunas)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Gagal memproses pengiriman: Kurir tidak tersedia."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/staff/orders": {
            "get": {
                "tags": [
                    "Orders (Staff)"
                ],
                "summary": "Daftar pesanan apotek (Staff)",
                "description": "Mengambil daftar pesanan yang masuk ke apotek tempat staf bekerja. Mendukung filter status dan paginasi.",
                "operationId": "staffListOrders",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter berdasarkan status pesanan (contoh: PENDING, PROCESSING, SHIPPED, COMPLETED, CANCELLED)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "PENDING"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Nomor halaman",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar pesanan berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Daftar pesanan apotek berhasil diambil"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Tidak terautentikasi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Staf tidak memiliki akses atau tidak aktif",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Anda tidak memiliki izin untuk mengakses apotek ini."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/staff/orders/{id}": {
            "get": {
                "tags": [
                    "Orders (Staff)"
                ],
                "summary": "Detail pesanan apotek (Staff)",
                "description": "Mengambil detail lengkap pesanan yang masuk ke apotek staf, termasuk item obat dan resep jika ada.",
                "operationId": "staffShowOrder",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID pesanan (UUID)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "019e2ac0-684e-7139-8cb9-c150e027c872"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail pesanan berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Detail pesanan berhasil diambil"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Tidak terautentikasi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pesanan tidak ditemukan di apotek ini",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Pesanan tidak ditemukan di apotek ini."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Masukkan token Sanctum Anda di sini. Token didapatkan dari endpoint POST /api/auth/login. Format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}